fix formatting and colors

+++ app/routes/clear.$deleteType.tsx
@@ -6,18 +6,13 @@ import { deleteAllItems, deletePurchasedItems } from '~/models/items'
 export const action = async ({ request, params }: ActionArgs) => {
   const user = await auth.isAuthenticated(request)
 
-  if (!params) {
-    throw new Error('Invalid data')
-  }
+  if (!params) throw new Error('Invalid data')
 
   if (user) {
-    if (params.deleteType === 'all') {
-      deleteAllItems(user.token)
-    }
+    if (params.deleteType === 'all') return deleteAllItems(user.token)
 
-    if (params.deleteType === 'purchased') {
-      deletePurchasedItems(user.token)
-    }
+    if (params.deleteType === 'purchased')
+      return deletePurchasedItems(user.token)
   } else {
     throw new Error('Please authenticate')
   }
